Rename _stats/histogram -> _stats/counting
