+++ seaborn/_stats/counting.py
