+++ seaborn/objects.py
@@ -37,7 +37,7 @@ from seaborn._marks.text import Text  # noqa: F401
 
 from seaborn._stats.base import Stat  # noqa: F401
 from seaborn._stats.aggregation import Agg, Est  # noqa: F401
-from seaborn._stats.histogram import Count, Hist  # noqa: F401
+from seaborn._stats.counting import Count, Hist  # noqa: F401
 from seaborn._stats.order import Perc  # noqa: F401
 from seaborn._stats.regression import PolyFit  # noqa: F401
